Finished tests

+++ src/main/java/com/jonrib/tasks/web/ResourceEntryCommentsController.java
@@ -48,7 +48,7 @@ public class ResourceEntryCommentsController {
 				return new ResponseEntity<String>("Resource entry not found", HttpStatus.NOT_FOUND);
 			}
 			if (!resourceEntryService.canEdit(entry.get(),DataController.getJWTCookie(request.getCookies())) && !resourceEntryService.canRead(entry.get(),DataController.getJWTCookie(request.getCookies()))) {
-				return new ResponseEntity<String>("Can't read resource entry", HttpStatus.BAD_REQUEST);
+				return new ResponseEntity<String>("Can't read resource entry", HttpStatus.FORBIDDEN);
 			}
 			return new ResponseEntity<String>(mapper.writeValueAsString(entry.get().getComments()), HttpStatus.OK);
 		}catch (Exception e) {
@@ -66,9 +66,6 @@ public class ResourceEntryCommentsController {
 			if (resEntry.isEmpty()) {
 				return new ResponseEntity<String>("Resource entry not found", HttpStatus.NOT_FOUND);
 			}
-			//if (!resourceEntryService.canEdit(resEntry.get(),DataController.getJWTCookie(request.getCookies())) && !resourceEntryService.canRead(resEntry.get(),DataController.getJWTCookie(request.getCookies()))) {
-				//return new ResponseEntity<String>("User can't edit resource entry", HttpStatus.BAD_REQUEST);
-			//}
 			User user = userService.findByUsername(securityService.findLoggedInUsername(DataController.getJWTCookie(request.getCookies())));
 			Set<Role> userRoles = user != null ? user.getRoles() : new HashSet<Role>();
 			boolean isAdmin = false;
@@ -78,24 +75,13 @@ public class ResourceEntryCommentsController {
 					break;
 				}
 			}
-			if (!isAdmin && !entry.get().getUserName().equals(securityService.findLoggedInUsername(DataController.getJWTCookie(request.getCookies()))))
-				return new ResponseEntity<String>("Only author or admin can delete comment", HttpStatus.BAD_REQUEST);
-			//removeFromComments(resEntry.get().getComments(), entry.get());
+			if (!isAdmin && (!securityService.findLoggedInUsername(DataController.getJWTCookie(request.getCookies())).equals("") || !entry.get().getUserName().equals(securityService.findLoggedInUsername(DataController.getJWTCookie(request.getCookies())))))
+				return new ResponseEntity<String>("Only author or admin can delete comment", HttpStatus.FORBIDDEN);
 			resourceEntryService.save(resEntry.get());
 			commentService.delete(entry.get());
 			return new ResponseEntity<String>("success", HttpStatus.OK);
 		}catch (Exception e) {
 			return new ResponseEntity<String>(e.toString(), HttpStatus.BAD_REQUEST);
-		}
-	}
-	
-	private void removeFromComments(Set<Comment> comments, Comment toRemove) {
-		if (comments == null)
-			return;
-		if (comments.remove(toRemove) == true) 
-			return;
-		for (Comment comm : comments) {
-			removeFromComments(comm.getReplies(), toRemove);
 		}
 	}
 	
@@ -106,9 +92,9 @@ public class ResourceEntryCommentsController {
 			if (resEntry.isEmpty()) {
 				return new ResponseEntity<String>("Resource entry not found", HttpStatus.NOT_FOUND);
 			}
-			//if (!resourceEntryService.canEdit(resEntry.get(),DataController.getJWTCookie(request.getCookies())) && !resourceEntryService.canRead(resEntry.get(),DataController.getJWTCookie(request.getCookies()))) {
-			//	return new ResponseEntity<String>("User can't edit resource entry", HttpStatus.BAD_REQUEST);
-			//}
+			if (!resourceEntryService.canEdit(resEntry.get(),DataController.getJWTCookie(request.getCookies())) && !resourceEntryService.canRead(resEntry.get(),DataController.getJWTCookie(request.getCookies()))) {
+				return new ResponseEntity<String>("Can't read resource entry", HttpStatus.FORBIDDEN);
+			}
 			
 			Comment entry = mapper.readValue(entryJson, Comment.class);
 			entry.setUserName(securityService.findLoggedInUsername(DataController.getJWTCookie(request.getCookies())));
@@ -118,7 +104,7 @@ public class ResourceEntryCommentsController {
 			resEntry.get().getComments().add(entry);
 			resourceEntryService.save(resEntry.get());
 			
-			return new ResponseEntity<String>(rez.getId()+"", HttpStatus.OK);
+			return new ResponseEntity<String>(mapper.writeValueAsString(rez), HttpStatus.OK);
 		}catch (Exception e) {
 			return new ResponseEntity<String>(e.toString(), HttpStatus.BAD_REQUEST);
 		}
@@ -134,21 +120,18 @@ public class ResourceEntryCommentsController {
 			if (resEntry.isEmpty()) {
 				return new ResponseEntity<String>("Resource entry not found", HttpStatus.NOT_FOUND);
 			}
-			//if (!resourceEntryService.canEdit(resEntry.get(),DataController.getJWTCookie(request.getCookies())) && !resourceEntryService.canRead(resEntry.get(),DataController.getJWTCookie(request.getCookies()))) {
-				//return new ResponseEntity<String>("User can't edit resource entry", HttpStatus.BAD_REQUEST);
-			//}
+			if (!resourceEntryService.canEdit(resEntry.get(),DataController.getJWTCookie(request.getCookies())) && !resourceEntryService.canRead(resEntry.get(),DataController.getJWTCookie(request.getCookies()))) {
+				return new ResponseEntity<String>("Can't read resource entry", HttpStatus.FORBIDDEN);
+			}
 			Comment reply = mapper.readValue(entryJson, Comment.class);
 			reply.setUserName(securityService.findLoggedInUsername(DataController.getJWTCookie(request.getCookies())));
 			reply.setDate(new Date());
-			//entry.get().setReplies(new HashSet<Comment>());
 			entry.get().getReplies().add(reply);
 			
-			//resEntry.get().getComments().add(entry);
-			//resourceEntryService.save(resEntry.get());
 			Comment rez = commentService.save(reply);
 			commentService.save(entry.get());
 			
-			return new ResponseEntity<String>(rez.getId()+"", HttpStatus.OK);
+			return new ResponseEntity<String>(mapper.writeValueAsString(rez), HttpStatus.OK);
 		}catch (Exception e) {
 			return new ResponseEntity<String>(e.toString(), HttpStatus.BAD_REQUEST);
 		}
